Use the new fake_probe module properly
